[test/fuzzing] Merge hb_fuzzer_tools.py back and simplify
